Migrate registry sync impl to use large packets from networking API
