+++ fabric-registry-sync-v0/src/client/java/net/fabricmc/fabric/impl/client/registry/sync/FabricRegistryClientInit.java
@@ -16,56 +16,18 @@
 
 package net.fabricmc.fabric.impl.client.registry.sync;
 
-import java.util.concurrent.CompletionException;
-
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import net.minecraft.text.Text;
-
 import net.fabricmc.api.ClientModInitializer;
 import net.fabricmc.fabric.api.client.networking.v1.ClientConfigurationNetworking;
-import net.fabricmc.fabric.impl.registry.sync.RegistrySyncManager;
-import net.fabricmc.fabric.impl.registry.sync.RemapException;
-import net.fabricmc.fabric.impl.registry.sync.SyncCompletePayload;
-import net.fabricmc.fabric.impl.registry.sync.packet.RegistryPacketHandler;
+import net.fabricmc.fabric.impl.registry.sync.packet.RegistrySyncPayload;
 
 public class FabricRegistryClientInit implements ClientModInitializer {
 	private static final Logger LOGGER = LoggerFactory.getLogger(FabricRegistryClientInit.class);
 
 	@Override
 	public void onInitializeClient() {
-		registerSyncPacketReceiver(RegistrySyncManager.DIRECT_PACKET_HANDLER);
-	}
-
-	private <T extends RegistryPacketHandler.RegistrySyncPayload> void registerSyncPacketReceiver(RegistryPacketHandler<T> packetHandler) {
-		ClientConfigurationNetworking.registerGlobalReceiver(packetHandler.getPacketId(), (payload, context) -> {
-			ClientRegistrySyncHandler.receivePacket(context.client(), packetHandler, payload, RegistrySyncManager.DEBUG || !context.client().isInSingleplayer())
-					.whenComplete((complete, throwable) -> {
-						if (throwable != null) {
-							LOGGER.error("Registry remapping failed!", throwable);
-							context.client().execute(() -> context.responseSender().disconnect(getText(throwable)));
-							return;
-						}
-
-						if (complete) {
-							context.responseSender().sendPacket(SyncCompletePayload.INSTANCE);
-						}
-					});
-		});
-	}
-
-	private Text getText(Throwable e) {
-		if (e instanceof RemapException remapException) {
-			final Text text = remapException.getText();
-
-			if (text != null) {
-				return text;
-			}
-		} else if (e instanceof CompletionException completionException) {
-			return getText(completionException.getCause());
-		}
-
-		return Text.literal("Registry remapping failed: " + e.getMessage());
+		ClientConfigurationNetworking.registerGlobalReceiver(RegistrySyncPayload.ID, ClientRegistrySyncHandler::receivePacket);
 	}
 }
