+++ fabric-registry-sync-v0/src/main/java/net/fabricmc/fabric/impl/registry/sync/packet/RegistrySyncPayload.java
@@ -1,0 +1,214 @@
+/*
+ * Copyright (c) 2016, 2017, 2018, 2019 FabricMC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package net.fabricmc.fabric.impl.registry.sync.packet;
+
+import java.util.ArrayList;
+import java.util.Comparator;
+import java.util.EnumSet;
+import java.util.Iterator;
+import java.util.LinkedHashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.stream.Collectors;
+
+import it.unimi.dsi.fastutil.objects.Object2IntLinkedOpenHashMap;
+import it.unimi.dsi.fastutil.objects.Object2IntMap;
+
+import net.minecraft.network.PacketByteBuf;
+import net.minecraft.network.codec.PacketCodec;
+import net.minecraft.network.packet.CustomPayload;
+import net.minecraft.registry.Registry;
+import net.minecraft.registry.RegistryKey;
+import net.minecraft.util.Identifier;
+
+import net.fabricmc.fabric.api.event.registry.RegistryAttribute;
+import net.fabricmc.fabric.api.event.registry.RegistryAttributeHolder;
+import net.fabricmc.fabric.impl.registry.sync.RegistryAttributeImpl;
+
+public record RegistrySyncPayload(
+		Map<Identifier, Object2IntMap<Identifier>> registryMap,
+		Map<Identifier, EnumSet<RegistryAttribute>> registryAttributes
+) implements CustomPayload {
+	public static CustomPayload.Id<RegistrySyncPayload> ID = new CustomPayload.Id<>(Identifier.of("fabric", "registry/sync"));
+	public static PacketCodec<PacketByteBuf, RegistrySyncPayload> CODEC = CustomPayload.codecOf(RegistrySyncPayload::write, RegistrySyncPayload::read);
+
+	public RegistrySyncPayload(Map<Identifier, Object2IntMap<Identifier>> registryMap) {
+		this(registryMap, getRegistryAttributeMap(registryMap));
+	}
+
+	private static Map<Identifier, EnumSet<RegistryAttribute>> getRegistryAttributeMap(Map<Identifier, Object2IntMap<Identifier>> registryMap) {
+		Map<Identifier, EnumSet<RegistryAttribute>> registryAttributes = new LinkedHashMap<>();
+		registryMap.forEach((regId, idMap) -> {
+			RegistryKey<Registry<Object>> registryKey = RegistryKey.ofRegistry(regId);
+			RegistryAttributeImpl holder = (RegistryAttributeImpl) RegistryAttributeHolder.get(registryKey);
+			registryAttributes.put(regId, holder.getAttributes());
+		});
+		return registryAttributes;
+	}
+
+	private static RegistrySyncPayload read(PacketByteBuf combinedBuf) {
+		Map<Identifier, Object2IntMap<Identifier>> syncedRegistryMap = new LinkedHashMap<>();
+		Map<Identifier, EnumSet<RegistryAttribute>> syncedRegistryAttributes = new LinkedHashMap<>();
+		int regNamespaceGroupAmount = combinedBuf.readVarInt();
+
+		for (int i = 0; i < regNamespaceGroupAmount; i++) {
+			String regNamespace = unoptimizeNamespace(combinedBuf.readString());
+			int regNamespaceGroupLength = combinedBuf.readVarInt();
+
+			for (int j = 0; j < regNamespaceGroupLength; j++) {
+				String regPath = combinedBuf.readString();
+				EnumSet<RegistryAttribute> attributes = decodeRegistryAttributes(combinedBuf.readByte());
+				Object2IntMap<Identifier> idMap = new Object2IntLinkedOpenHashMap<>();
+				int idNamespaceGroupAmount = combinedBuf.readVarInt();
+
+				int lastBulkLastRawId = 0;
+
+				for (int k = 0; k < idNamespaceGroupAmount; k++) {
+					String idNamespace = unoptimizeNamespace(combinedBuf.readString());
+					int rawIdBulkAmount = combinedBuf.readVarInt();
+
+					for (int l = 0; l < rawIdBulkAmount; l++) {
+						int bulkRawIdStartDiff = combinedBuf.readVarInt();
+						int bulkSize = combinedBuf.readVarInt();
+
+						int currentRawId = (lastBulkLastRawId + bulkRawIdStartDiff) - 1;
+
+						for (int m = 0; m < bulkSize; m++) {
+							currentRawId++;
+							String idPath = combinedBuf.readString();
+							idMap.put(Identifier.of(idNamespace, idPath), currentRawId);
+						}
+
+						lastBulkLastRawId = currentRawId;
+					}
+				}
+
+				Identifier registryId = Identifier.of(regNamespace, regPath);
+				syncedRegistryMap.put(registryId, idMap);
+				syncedRegistryAttributes.put(registryId, attributes);
+			}
+		}
+
+		return new RegistrySyncPayload(syncedRegistryMap, syncedRegistryAttributes);
+	}
+
+	private void write(PacketByteBuf buf) {
+		// Group registry ids with same namespace.
+		Map<String, List<Identifier>> regNamespaceGroups = registryMap.keySet().stream()
+				.collect(Collectors.groupingBy(Identifier::getNamespace));
+
+		buf.writeVarInt(regNamespaceGroups.size());
+
+		regNamespaceGroups.forEach((regNamespace, regIds) -> {
+			buf.writeString(optimizeNamespace(regNamespace));
+			buf.writeVarInt(regIds.size());
+
+			for (Identifier regId : regIds) {
+				buf.writeString(regId.getPath());
+				buf.writeByte(encodeRegistryAttributes(registryAttributes.getOrDefault(regId, EnumSet.noneOf(RegistryAttribute.class))));
+
+				Object2IntMap<Identifier> idMap = registryMap.get(regId);
+
+				// Sort object ids by its namespace. We use linked map here to keep the original namespace ordering.
+				Map<String, List<Object2IntMap.Entry<Identifier>>> idNamespaceGroups = idMap.object2IntEntrySet().stream()
+						.collect(Collectors.groupingBy(e -> e.getKey().getNamespace(), LinkedHashMap::new, Collectors.toCollection(ArrayList::new)));
+
+				buf.writeVarInt(idNamespaceGroups.size());
+
+				int lastBulkLastRawId = 0;
+
+				for (Map.Entry<String, List<Object2IntMap.Entry<Identifier>>> idNamespaceEntry : idNamespaceGroups.entrySet()) {
+					// Make sure the ids are sorted by its raw id.
+					List<Object2IntMap.Entry<Identifier>> idPairs = idNamespaceEntry.getValue();
+					idPairs.sort(Comparator.comparingInt(Object2IntMap.Entry::getIntValue));
+
+					// Group consecutive raw ids together.
+					List<List<Object2IntMap.Entry<Identifier>>> bulks = new ArrayList<>();
+
+					Iterator<Object2IntMap.Entry<Identifier>> idPairIter = idPairs.iterator();
+					List<Object2IntMap.Entry<Identifier>> currentBulk = new ArrayList<>();
+					Object2IntMap.Entry<Identifier> currentPair = idPairIter.next();
+					currentBulk.add(currentPair);
+
+					while (idPairIter.hasNext()) {
+						currentPair = idPairIter.next();
+
+						if (currentBulk.get(currentBulk.size() - 1).getIntValue() + 1 != currentPair.getIntValue()) {
+							bulks.add(currentBulk);
+							currentBulk = new ArrayList<>();
+						}
+
+						currentBulk.add(currentPair);
+					}
+
+					bulks.add(currentBulk);
+
+					buf.writeString(optimizeNamespace(idNamespaceEntry.getKey()));
+					buf.writeVarInt(bulks.size());
+
+					for (List<Object2IntMap.Entry<Identifier>> bulk : bulks) {
+						int firstRawId = bulk.get(0).getIntValue();
+						int bulkRawIdStartDiff = firstRawId - lastBulkLastRawId;
+
+						buf.writeVarInt(bulkRawIdStartDiff);
+						buf.writeVarInt(bulk.size());
+
+						for (Object2IntMap.Entry<Identifier> idPair : bulk) {
+							buf.writeString(idPair.getKey().getPath());
+
+							lastBulkLastRawId = idPair.getIntValue();
+						}
+					}
+				}
+			}
+		});
+	}
+
+	private static byte encodeRegistryAttributes(EnumSet<RegistryAttribute> attributes) {
+		byte encoded = 0;
+
+		// Only send the optional marker.
+		if (attributes.contains(RegistryAttribute.OPTIONAL)) {
+			encoded |= 0x1;
+		}
+
+		return encoded;
+	}
+
+	private static EnumSet<RegistryAttribute> decodeRegistryAttributes(byte encoded) {
+		EnumSet<RegistryAttribute> attributes = EnumSet.noneOf(RegistryAttribute.class);
+
+		if ((encoded & 0x1) != 0) {
+			attributes.add(RegistryAttribute.OPTIONAL);
+		}
+
+		return attributes;
+	}
+
+	private static String optimizeNamespace(String namespace) {
+		return namespace.equals(Identifier.DEFAULT_NAMESPACE) ? "" : namespace;
+	}
+
+	private static String unoptimizeNamespace(String namespace) {
+		return namespace.isEmpty() ? Identifier.DEFAULT_NAMESPACE : namespace;
+	}
+
+	@Override
+	public Id<RegistrySyncPayload> getId() {
+		return ID;
+	}
+}
